pyproject.toml
[tool.poetry]
name = "sinotrack-alert-monitor"
version = "0.1.0"
description = ""
authors = ["kpliuta <9212935+kpliuta@users.noreply.github.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.21.0"
requests = "^2.32.5"
python-dotenv = "^1.1.1"
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
